Corrected all, except 'private'
