+++ src/main/java/data/provider/LoginUser.java
@@ -1,4 +1,4 @@
-package dataProvider;
+package data.provider;
 
 public class LoginUser {
     private String email;
