+++ src/test/java/pageObjects/Constructor.java
@@ -1,13 +1,13 @@
 package pageObjects;
 
-import basePages.BasePage;
+import base.pages.BasePage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 public class Constructor extends BasePage {
-    public static final By BUN_BUTTON = By.xpath("//*[@id=\"root\"]/div/main/section[1]/div[1]/div[1]");
-    public static final By SOUCES_BUTTON = By.xpath("//*[@id=\"root\"]/div/main/section[1]/div[1]/div[2]");
-    public static final By FILLERS_BUTTON = By.xpath("//*[@id=\"root\"]/div/main/section[1]/div[1]/div[3]");
+    public static final By BUN_BUTTON = By.cssSelector("div.tab_tab__1SPyG:nth-child(1)");
+    public static final By SOUCES_BUTTON = By.cssSelector("div.tab_tab__1SPyG:nth-child(2)");
+    public static final By FILLERS_BUTTON = By.cssSelector("div.tab_tab__1SPyG:nth-child(3)");
     public Constructor(WebDriver driver) {
         super(driver);
     }
