+++ src/test/java/pageObjects/ProfilePage.java
@@ -1,11 +1,11 @@
 package pageObjects;
 
-import basePages.BasePage;
+import base.pages.BasePage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 public class ProfilePage extends BasePage {
-    public static final By LOGOUT_BUTTON_IN_PROFILE = By.xpath("//*[@id=\"root\"]/div/main/div/nav/ul/li[3]/button");
+    public static final By LOGOUT_BUTTON_IN_PROFILE = By.cssSelector(".Account_button__14Yp3");
 
     public ProfilePage(WebDriver driver) {
         super(driver);
