+++ src/test/java/tests/enter/account/EnterAccountFromEnterButtonInForgotPasswordForm.java
@@ -1,9 +1,9 @@
-package tests.enterPersonalAccountTests;
+package tests.enter.account;
 
-import basePages.ConfigBrowser;
+import base.pages.ConfigBrowser;
 import clients.UserClient;
-import dataProvider.CreateUser;
-import dataProvider.RegFormRandomData;
+import data.provider.CreateUser;
+import data.provider.RegFormRandomData;
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Assert;
@@ -12,24 +12,25 @@ import org.openqa.selenium.WebDriver;
 import pageObjects.AuthFormPage;
 
 import static constants.LocatorsAndDataConstants.CHECKOUT_ORDER_BUTTON;
-import static pageObjects.RegistrationFormPage.clickEnterButtonInRegForm;
-import static pageObjects.RegistrationFormPage.openRegistrationPage;
+import static pageObjects.AuthFormPage.authorize;
+import static pageObjects.ForgotPasswordPage.clickEnterButtonInFogrotPasswordForm;
+import static pageObjects.ForgotPasswordPage.openForgotPasswordPage;
 
-public class EnterAccountWithEnterButtonInRegFormTest {
+public class EnterAccountFromEnterButtonInForgotPasswordForm {
     public static String accessToken;
     WebDriver driver = ConfigBrowser.startDriver();
     AuthFormPage authFormPage = new AuthFormPage(driver);
 
     @Test
-    @DisplayName("Вход в аккаунт через кнопку Войти в форме регистрации")
+    @DisplayName("Вход в личный кабинет из формы 'Забыл пароль'")
     public void enterPersonalAccountTest() {
 
         CreateUser createUser = RegFormRandomData.getUserData();
         accessToken = UserClient.create(createUser).extract().jsonPath().get("accessToken");
 
-        openRegistrationPage();
-        clickEnterButtonInRegForm();
-        AuthFormPage.authorize(createUser);
+        openForgotPasswordPage();
+        clickEnterButtonInFogrotPasswordForm();
+        authorize(createUser);
 
         boolean checkoutOrderButton = driver.findElement(CHECKOUT_ORDER_BUTTON).isDisplayed();
         Assert.assertTrue(checkoutOrderButton);
