+++ src/test/java/tests/enter/account/EnterPersonalAccountTest.java
@@ -1,35 +1,35 @@
-package tests.enterPersonalAccountTests;
+package tests.enter.account;
 
-import basePages.ConfigBrowser;
+import base.pages.ConfigBrowser;
 import clients.UserClient;
-import dataProvider.CreateUser;
-import dataProvider.RegFormRandomData;
+import data.provider.CreateUser;
+import data.provider.RegFormRandomData;
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
 import pageObjects.AuthFormPage;
+import pageObjects.BurgerHomePage;
 
+import static constants.ConstantUrls.STELLAR_BURGERS_HOME_PAGE;
 import static constants.LocatorsAndDataConstants.CHECKOUT_ORDER_BUTTON;
 import static pageObjects.AuthFormPage.authorize;
-import static pageObjects.ForgotPasswordPage.clickEnterButtonInFogrotPasswordForm;
-import static pageObjects.ForgotPasswordPage.openForgotPasswordPage;
 
-public class EnterAccountFromEnterButtonInForgotPasswordForm {
+public class EnterPersonalAccountTest {
     public static String accessToken;
     WebDriver driver = ConfigBrowser.startDriver();
     AuthFormPage authFormPage = new AuthFormPage(driver);
 
     @Test
-    @DisplayName("Вход в личный кабинет из формы 'Забыл пароль'")
+    @DisplayName("Вход по кнопке «Войти в аккаунт» на главной")
     public void enterPersonalAccountTest() {
 
         CreateUser createUser = RegFormRandomData.getUserData();
         accessToken = UserClient.create(createUser).extract().jsonPath().get("accessToken");
 
-        openForgotPasswordPage();
-        clickEnterButtonInFogrotPasswordForm();
+        driver.get(STELLAR_BURGERS_HOME_PAGE);
+        BurgerHomePage.enterAccountButtonClick();
         authorize(createUser);
 
         boolean checkoutOrderButton = driver.findElement(CHECKOUT_ORDER_BUTTON).isDisplayed();
