+++ src/test/java/tests/profile/EnterProfileWithHeaderProfileButtonTest.java
@@ -1,9 +1,9 @@
-package tests.ProfileTests;
+package tests.profile;
 
-import basePages.ConfigBrowser;
+import base.pages.ConfigBrowser;
 import clients.UserClient;
-import dataProvider.CreateUser;
-import dataProvider.RegFormRandomData;
+import data.provider.CreateUser;
+import data.provider.RegFormRandomData;
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Assert;
@@ -14,32 +14,31 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 import pageObjects.AuthFormPage;
 
 import static constants.ConstantUrls.PROFILE_PAGE;
-import static constants.LocatorsAndDataConstants.CHECKOUT_ORDER_BUTTON;
 import static pageObjects.AuthFormPage.openLoginPage;
-import static pageObjects.HeaderButtons.clickHeaderConstructorButton;
 import static pageObjects.HeaderButtons.clickProfileButton;
+import static pageObjects.RegistrationFormPage.openRegistrationPage;
+import static pageObjects.RegistrationFormPage.registerUser;
 
-public class ConstructorLinkFromProfileTest {
+public class EnterProfileWithHeaderProfileButtonTest {
     public static String accessToken;
     WebDriver driver = ConfigBrowser.startDriver();
     AuthFormPage authFormPage = new AuthFormPage(driver);
 
     @Test
-    @DisplayName("Переход из личного кабинета в конструктор")
+    @DisplayName("Переход по клику на «Личный кабинет»")
     public void enterPersonalAccountTest() {
 
         CreateUser createUser = RegFormRandomData.getUserData();
         accessToken = UserClient.create(createUser).extract().jsonPath().get("accessToken");
 
+        openRegistrationPage();
+        registerUser(createUser);
         openLoginPage();
         AuthFormPage.authorize(createUser);
         clickProfileButton();
         new WebDriverWait(driver, 5).until(ExpectedConditions.urlToBe(PROFILE_PAGE));
-        clickHeaderConstructorButton();
 
-        boolean checkoutOrderButton = driver.findElement(CHECKOUT_ORDER_BUTTON).isDisplayed();
-        Assert.assertTrue(checkoutOrderButton);
-
+        Assert.assertEquals(PROFILE_PAGE, driver.getCurrentUrl());
     }
 
     @After
